Split client endpoints into modules
